koordlet: cleanup cgroup path utils
