+++ pkg/koordlet/util/meminfo_test.go
@@ -19,51 +19,116 @@ package util
 import (
 	"os"
 	"path/filepath"
-	"runtime"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/koordinator-sh/koordinator/pkg/koordlet/util/system"
 )
 
 func Test_readMemInfo(t *testing.T) {
 	tempDir := t.TempDir()
 	tempInvalidMemInfoPath := filepath.Join(tempDir, "no_meminfo")
 	tempMemInfoPath := filepath.Join(tempDir, "meminfo")
-	memInfoContentStr := "MemTotal:       263432804 kB\nMemFree:        254391744 kB\nMemAvailable:   256703236 kB\n" +
-		"Buffers:          958096 kB\nCached:          3763224 kB\nSwapCached:            0 kB\n" +
-		"Active:          2786012 kB\nInactive:        2223752 kB\nActive(anon):     289488 kB\n" +
-		"Inactive(anon):     1300 kB\nActive(file):    2496524 kB\nInactive(file):  2222452 kB\n" +
-		"Unevictable:           0 kB\nMlocked:               0 kB\nSwapTotal:             0 kB\n" +
-		"SwapFree:              0 kB\nDirty:               624 kB\nWriteback:             0 kB\n" +
-		"AnonPages:        281748 kB\nMapped:           495936 kB\nShmem:              2340 kB\n" +
-		"Slab:            1097040 kB\nSReclaimable:     445164 kB\nSUnreclaim:       651876 kB\n" +
-		"KernelStack:       20944 kB\nPageTables:         7896 kB\nNFS_Unstable:          0 kB\n" +
-		"Bounce:                0 kB\nWritebackTmp:          0 kB\nCommitLimit:    131716400 kB\n" +
-		"Committed_AS:    3825364 kB\nVmallocTotal:   34359738367 kB\nVmallocUsed:           0 kB\n" +
-		"VmallocChunk:          0 kB\nHardwareCorrupted:     0 kB\nAnonHugePages:     38912 kB\n" +
-		"ShmemHugePages:        0 kB\nShmemPmdMapped:        0 kB\nCmaTotal:              0 kB\n" +
-		"CmaFree:               0 kB\nHugePages_Total:       0\nHugePages_Free:        0\n" +
-		"HugePages_Rsvd:        0\nHugePages_Surp:        0\nHugepagesize:       2048 kB\n" +
-		"DirectMap4k:      414760 kB\nDirectMap2M:     8876032 kB\nDirectMap1G:    261095424 kB\n"
+	memInfoContentStr := `MemTotal:       263432804 kB
+MemFree:        254391744 kB
+MemAvailable:   256703236 kB
+Buffers:          958096 kB
+Cached:          3763224 kB
+SwapCached:            0 kB
+Active:          2786012 kB
+Inactive:        2223752 kB
+Active(anon):     289488 kB
+Inactive(anon):     1300 kB
+Active(file):    2496524 kB
+Inactive(file):  2222452 kB
+Unevictable:           0 kB
+Mlocked:               0 kB
+SwapTotal:             0 kB
+SwapFree:              0 kB
+Dirty:               624 kB
+Writeback:             0 kB
+AnonPages:        281748 kB
+Mapped:           495936 kB
+Shmem:              2340 kB
+Slab:            1097040 kB
+SReclaimable:     445164 kB
+SUnreclaim:       651876 kB
+KernelStack:       20944 kB
+PageTables:         7896 kB
+NFS_Unstable:          0 kB
+Bounce:                0 kB
+WritebackTmp:          0 kB
+CommitLimit:    131716400 kB
+Committed_AS:    3825364 kB
+VmallocTotal:   34359738367 kB
+VmallocUsed:           0 kB
+VmallocChunk:          0 kB
+HardwareCorrupted:     0 kB
+AnonHugePages:     38912 kB
+ShmemHugePages:        0 kB
+ShmemPmdMapped:        0 kB
+CmaTotal:              0 kB
+CmaFree:               0 kB
+HugePages_Total:       0
+HugePages_Free:        0
+HugePages_Rsvd:        0
+HugePages_Surp:        0
+Hugepagesize:       2048 kB
+DirectMap4k:      414760 kB
+DirectMap2M:     8876032 kB
+DirectMap1G:    261095424 kB`
 	err := os.WriteFile(tempMemInfoPath, []byte(memInfoContentStr), 0666)
 	assert.NoError(t, err)
 	tempMemInfoPath1 := filepath.Join(tempDir, "meminfo1")
-	memInfoContentStr1 := "MemTotal:       263432804 kB\nMemFree:        254391744 kB\nMemAvailable:   256703236 kB\n" +
-		"Buffers:          958096 kB\nCached:          invalidField kB\nSwapCached:            0 kB\n" +
-		"Active:          2786012 kB\nInactive:        2223752 kB\nActive(anon):     289488 kB\n" +
-		"Inactive(anon):     1300 kB\nActive(file):    2496524 kB\nInactive(file):  2222452 kB\n" +
-		"Unevictable:           0 kB\nMlocked:               0 kB\nSwapTotal:             0 kB\n" +
-		"SwapFree:              0 kB\nDirty:               624 kB\nWriteback:             0 kB\n" +
-		"AnonPages:        281748 kB\nMapped:           495936 kB\nShmem:              2340 kB\n" +
-		"Slab:            1097040 kB\nSReclaimable:     445164 kB\nSUnreclaim:       651876 kB\n" +
-		"KernelStack:       20944 kB\nPageTables:         7896 kB\nNFS_Unstable:          0 kB\n" +
-		"Bounce:                0 kB\nWritebackTmp:          0 kB\nCommitLimit:    131716400 kB\n" +
-		"Committed_AS:    3825364 kB\nVmallocTotal:   34359738367 kB\nVmallocUsed:           0 kB\n" +
-		"VmallocChunk:          0 kB\nHardwareCorrupted:     0 kB\nAnonHugePages:     38912 kB\n" +
-		"ShmemHugePages:        0 kB\nShmemPmdMapped:        0 kB\nCmaTotal:              0 kB\n" +
-		"CmaFree:               0 kB\nHugePages_Total:       0\nHugePages_Free:        0\n" +
-		"HugePages_Rsvd:        0\nHugePages_Surp:        0\nHugepagesize:       2048 kB\n" +
-		"DirectMap4k:      414760 kB\nDirectMap2M:     8876032 kB\nDirectMap1G:    261095424 kB\n"
+	memInfoContentStr1 := `MemTotal:       263432804 kB
+MemFree:        254391744 kB
+MemAvailable:   256703236 kB
+Buffers:          958096 kB
+Cached:                0 kB
+SwapCached:            0 kB
+Active:          2786012 kB
+Inactive:        2223752 kB
+Active(anon):     289488 kB
+Inactive(anon):     1300 kB
+Active(file):    2496524 kB
+Inactive(file):  2222452 kB
+Unevictable:           0 kB
+Mlocked:               0 kB
+SwapTotal:             0 kB
+SwapFree:              0 kB
+Dirty:               624 kB
+Writeback:             0 kB
+AnonPages:        281748 kB
+Mapped:           495936 kB
+Shmem:              2340 kB
+Slab:            1097040 kB
+SReclaimable:     445164 kB
+SUnreclaim:       651876 kB
+KernelStack:       20944 kB
+PageTables:         7896 kB
+NFS_Unstable:          0 kB
+Bounce:                0 kB
+WritebackTmp:          0 kB
+CommitLimit:    131716400 kB
+Committed_AS:    3825364 kB
+VmallocTotal:   34359738367 kB
+VmallocUsed:           0 kB
+VmallocChunk:          0 kB
+HardwareCorrupted:     0 kB
+AnonHugePages:     38912 kB
+ShmemHugePages:        0 kB
+ShmemPmdMapped:        0 kB
+CmaTotal:              0 kB
+CmaFree:               0 kB
+HugePages_Total:       0
+HugePages_Free:        0
+HugePages_Rsvd:        0
+HugePages_Surp:        0
+Hugepagesize:       2048 kB
+DirectMap4k:      414760 kB
+DirectMap2M:     8876032 kB
+DirectMap1G:    261095424 kB`
 	err = os.WriteFile(tempMemInfoPath1, []byte(memInfoContentStr1), 0666)
 	assert.NoError(t, err)
 	type args struct {
@@ -150,25 +215,68 @@ func Test_readMemInfo(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := readMemInfo(tt.args.path)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("readMemInfo wantErr %v but got err %s", tt.wantErr, err)
-			}
-			if !tt.wantErr {
-				assert.Equal(t, tt.want, got)
-			}
+			got, gotErr := readMemInfo(tt.args.path)
+			assert.Equal(t, tt.wantErr, gotErr != nil)
+			assert.Equal(t, tt.want, got)
 		})
 	}
 }
 
 func Test_GetMemInfoUsageKB(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Log("Ignore non-Linux environment")
-		return
-	}
+	testMemInfo := `MemTotal:       263432804 kB
+MemFree:        254391744 kB
+MemAvailable:   256703236 kB
+Buffers:          958096 kB
+Cached:          3763224 kB
+SwapCached:            0 kB
+Active:          2786012 kB
+Inactive:        2223752 kB
+Active(anon):     289488 kB
+Inactive(anon):     1300 kB
+Active(file):    2496524 kB
+Inactive(file):  2222452 kB
+Unevictable:           0 kB
+Mlocked:               0 kB
+SwapTotal:             0 kB
+SwapFree:              0 kB
+Dirty:               624 kB
+Writeback:             0 kB
+AnonPages:        281748 kB
+Mapped:           495936 kB
+Shmem:              2340 kB
+Slab:            1097040 kB
+SReclaimable:     445164 kB
+SUnreclaim:       651876 kB
+KernelStack:       20944 kB
+PageTables:         7896 kB
+NFS_Unstable:          0 kB
+Bounce:                0 kB
+WritebackTmp:          0 kB
+CommitLimit:    131716400 kB
+Committed_AS:    3825364 kB
+VmallocTotal:   34359738367 kB
+VmallocUsed:           0 kB
+VmallocChunk:          0 kB
+HardwareCorrupted:     0 kB
+AnonHugePages:     38912 kB
+ShmemHugePages:        0 kB
+ShmemPmdMapped:        0 kB
+CmaTotal:              0 kB
+CmaFree:               0 kB
+HugePages_Total:       0
+HugePages_Free:        0
+HugePages_Rsvd:        0
+HugePages_Surp:        0
+Hugepagesize:       2048 kB
+DirectMap4k:      414760 kB
+DirectMap2M:     8876032 kB
+DirectMap1G:    261095424 kB`
+
+	helper := system.NewFileTestUtil(t)
+	defer helper.Cleanup()
+	helper.WriteProcSubFileContents(system.ProcMemInfoName, testMemInfo)
+
 	memInfoUsage, err := GetMemInfoUsageKB()
-	if err != nil {
-		t.Error("failed to get MemInfo usage: ", err)
-	}
-	t.Log("meminfo: ", memInfoUsage)
+	assert.NoError(t, err)
+	assert.NotNil(t, memInfoUsage)
 }
